video_processing/video_transcription_pipeline/requirements.txt
openai-whisper>=20231117
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.8.1
soundfile>=0.10.3
ffmpeg-python>=0.2.0
moviepy>=1.0.3
openai>=1.0.0
pyyaml>=6.0